Fix some obscure bugs.
ms-to-uvfits wasn't working on garrawarla (something to do with
"relocation"); statically compiling ERFA with -fPIE works, so now using
a newer version of erfa-sys with that fix.

While testing the new ms-to-uvfits, I got some unusual "too many I/O
drivers" error messages from cfitsio. I'm not sure if that coincides
with some other activity on the supercomputer. In any case, turning off
parallelism when creating uvfits files seems to fix this, and this
change doesn't sacrifice much runtime performance. I put a progress bar
in, which is better for the user to see that the program is busy, anyway.